Always prefer `window` over `this` for global object detection.
Depending on how this module is consumed by applications, it might
be wrapped by a module-pattern function. In this case `this` no
longer refers to the global object, but to the module object, so
the polyfill fails.

This change checks if `window` is present. If so, use that, since
in this case, that definitely always refers to the global. To
maintain compatibility (with node, or other non-browser envs),
the existing `this` is used as a fallback.